server_py/requirements.txt
fastapi==0.115.12
python-socketio==5.13.0
uvicorn==0.34.0
